Updated 3dpose for sending the coordinates to Unity, plus the Unity c# file that receives/uses the data.
As I worried, this doesn't really work because the coordinates OpenCV is sending are in camera space. The camera is looking down, for example, so its z axis it's reporting a z coordinate for is rotated into its viewing angle. So I'm back to needing a way to calculate the camera's transform again. :/